requirements.txt
streamlit==1.42.0
pytesseract==0.3.10
pdfplumber==0.11.0
google-generativeai==0.8.4
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.0.1
pydantic==2.10.6
